# services/auth/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  description: Сервис аутентификации и управления пользователями
  version: 1.0.0
  contact:
    name: Auth Service Team

servers:
  - url: http://localhost:8081
    description: Локальный сервер
  - url: http://auth-service:8081
    description: Docker контейнер

tags:
  - name: Auth
    description: Аутентификация и авторизация
  - name: System
    description: Системные эндпоинты

paths:
  /health:
    get:
      tags:
        - System
      summary: Health check
      description: Проверка работоспособности сервиса
      operationId: healthCheck
      responses:
        '200':
          description: Сервис работает
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация пользователя
      description: Создание нового пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email уже зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags:
        - Auth
      summary: Вход пользователя
      description: Аутентификация пользователя и получение токенов
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      tags:
        - Auth
      summary: Информация о текущем пользователе
      description: Получение информации о текущем аутентифицированном пользователе
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Отсутствует заголовок Authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверный или отозванный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/validate:
    post:
      tags:
        - Auth
      summary: Проверка токена (внутренний endpoint)
      description: Проверка валидности access токена. Используется другими сервисами (Core, Collector)
      operationId: validateToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Токен валиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
        '400':
          description: Отсутствует заголовок Authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Токен невалиден или отозван
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      tags:
        - Auth
      summary: Выход из системы
      description: Отзыв access и refresh токенов (демонстрирует работу с blacklist)
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '400':
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен в формате "Bearer <token>"

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
          description: Email пользователя
        password:
          type: string
          minLength: 6
          example: password123
          description: Пароль (минимум 6 символов)

    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User created successfully
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: password123

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: Access токен (действует 15 минут)
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: Refresh токен (действует 7 дней)
        expires_in:
          type: integer
          example: 900
          description: Время жизни access токена в секундах
        token_type:
          type: string
          example: Bearer

    UserInfo:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        created_at:
          type: string
          format: date-time
          example: 2025-12-02T13:00:00Z

    ValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com

    LogoutRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: Refresh токен для отзыва

    LogoutResponse:
      type: object
      properties:
        message:
          type: string
          example: Logged out successfully

    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request body
          description: Описание ошибки